remove legacy chat importer (#75)
ChatVault is a rewrite of a[nother app ](https://github.com/vitormarcal/whatsapp-archiver) created in Node. That's why I created some APIs to facilitate the migration from the old one to this one.
As I was basically the only user of the old app, I am deleting the dead code since I no longer need this functionality.